# perf: add alternator frontend to perf_simple_query