@psarna 2021-05-17
v2:
 - abstracted out `alternator_test_env` helper class to a separate commit
 - added lots of warnings **not** to use the env for actual tests unless absolutely necessary