@psarna 2021-05-18
v2:
 - reduced some header noise by forward-declaring classes