@psarna 2021-05-18
v3:
 - fixed a configure.py issue, I was to eager to link alternator_test_env to unrelated objects