@nyh 2021-05-19
I'm too late in reviewing this, but as you can probably guess I have a philosophical disagreement with the new "alternator testenv" framework, which is a second and simplified copy of the regular Alternator initialization in main(), and will have all the usual problems and annoyance of code duplication - from now on, every time we change something in Alternator's internals we risk breaking this second initialization without even noticing (because I never run perf_simple_query as part of Alternator development). Beyond just duplicate initialization - the new code calls directly functions like `executor.update_item()` which until now were completely internal to Alternator - nothing outside the alternator/ directory included `executor.hh`. Now this changes. 

Personally, what I would have preferred to see instead is perf_simple_query starting a full-blown Scylla and not try to hack a second intialization codepath (or a third, after cql-repl). Also we could use the normal, external, API, to make requests instead of executor functions - or if we must use functions and not a socket, perhaps at least we should use a single function from server.cc, not the various functions from executor.cc.

Finally, another small rant (sorry): The nice long explanation of how to use perf_simple_query with alternator belongs in a file in docs/alternator - not in a commit message. Nobody will ever know how to find it in the commit message, and a month from now, even if I want to use perf_simple_query with Alternator I won't remember how.