@nyh 2021-05-19
> Personally, what I would have preferred to see instead is perf_simple_query starting a full-blown Scylla and not try to hack a second intialization codepath (or a third, after cql-repl). Also we could use the normal, external, API, to make requests instead of executor functions 

Expanding on this thought:

Another problem with the current approach (of a specialized `*_test_env`) is that it can't capture complexities like multi-node clusters and how those effect the comparisons you're trying to measure. For example, you reported a 3-fold slowdown with LWT. But this was on just one node... LWT probably becomes even slower if you have multiple nodes... So it sounds to me like the right thing to do would be to start a full-blown Scylla - one or more nodes - with some statistics turned on and some loader sending it requests. We could calculate on a real Scylla the average number of allocations per request, or the average number of instructions per request - two things you reported. What will become not interesting in such a setup is the real time measurements (TPS), etc., but the other counters will be just as valid even if you run multiple Scylla nodes on the same laptop.